t13: largeintbuffer in datamanager
zeiten sind aber durch auslagerung verfälscht, lief hart an der grenze
                        PC: Sven-LAPTOP
    Beginne Sortierung von: 1.000.000.000 Integers (3814MB)
Initial Integers pro Block: 640.679.936
   Integers pro Merge-Read: 33.554.432
  Integers pro Merge-Write: 33.554.432
       Max Arbeitsspeicher: 2.444MB
 Max GrÃ¶ÃŸe eines Readcalls: 128MB

readcall: 48588 ms
Start Quicksort
Stop Quicksort
quicksort: 115722 ms
writecall: 46666 ms
readcall: 42363 ms
Start Quicksort
Stop Quicksort
quicksort: 62415 ms
writecall: 50029 ms
readcall: 16 ms
06:22 - Diff 382.9s: QuickSort abgschlossen. Beginne ersten Merge-Schrit mit RunlÃ¤nge = 640.679.936
09:25 - Diff 182.0s: Sortieren abgeschlossen! Die Ausgabedatei enthÃ¤lt 1.000.000.000 Integers.

 ----- hier nochmal der normale buffer ----------

                        PC: Sven-LAPTOP
    Beginne Sortierung von: 1.000.000.000 Integers (3814MB)
Initial Integers pro Block: 500.000.001
   Integers pro Merge-Read: 33.554.432
  Integers pro Merge-Write: 33.554.432
       Max Arbeitsspeicher: 2.047MB
 Max GrÃ¶ÃŸe eines Readcalls: 128MB

readcall: 37990 ms
Start Quicksort
Stop Quicksort
quicksort: 46542 ms
writecall: 34120 ms
readcall: 34935 ms
Start Quicksort
Stop Quicksort
quicksort: 44739 ms
writecall: 33635 ms
readcall: 0 ms
03:53 - Diff 233.5s: QuickSort abgschlossen. Beginne ersten Merge-Schrit mit RunlÃ¤nge = 500.000.001
06:23 - Diff 150.2s: Sortieren abgeschlossen! Die Ausgabedatei enthÃ¤lt 1.000.000.000 Integers.

+++ ad-praktikum-t13-t14/src/aufgabe2/data/TestFileGenerator.java
@@ -82,7 +82,7 @@ public class TestFileGenerator {
     public static boolean isSorted(String fileName){
         Reader reader = Reader.create(fileName, Constants.BUFFERSIZE_MERGEREAD);
         long CONTROLL_COUNTER = 0;
-        ByteBuffer bbuffer = ByteBuffer.allocate(Constants.BUFFERSIZE_MERGEREAD);
+        ByteBuffer bbuffer = ByteBuffer.allocate((int)Constants.BUFFERSIZE_MERGEREAD);
        
         int overhangElem = Integer.MIN_VALUE;
